+++ es/diff_suppress_funcs.go
@@ -149,3 +149,14 @@ func diffSuppressPolicy(k, old, new string, d *schema.ResourceData) bool {
 
 	return reflect.DeepEqual(oo, no)
 }
+
+func diffSuppressLicense(k, old, new string, d *schema.ResourceData) bool {
+	var oldObj, newObj map[string]interface{}
+	if err := json.Unmarshal([]byte(old), &oldObj); err != nil {
+		return false
+	}
+	if err := json.Unmarshal([]byte(new), &newObj); err != nil {
+		return false
+	}
+	return reflect.DeepEqual(oldObj, newObj)
+}